Fix build issues

+++ Client/mods/deathmatch/logic/CClientVectorGraphicDisplay.cpp
@@ -102,7 +102,7 @@ void CClientVectorGraphicDisplay::UpdateTexture()
         auto surfaceData = static_cast<std::uint8_t*>(LockedRect.pBits);
         auto stride = static_cast<std::uint32_t>(LockedRect.Pitch);
 
-        Bitmap bitmap{surfaceData, pVectorGraphicItem->m_uiSizeX, pVectorGraphicItem->m_uiSizeY, stride};
+        Bitmap bitmap{surfaceData, (int32_t)pVectorGraphicItem->m_uiSizeX, (int32_t)pVectorGraphicItem->m_uiSizeY, (int32_t)stride};
         bitmap.clear(0);
         svgDocument->render(bitmap, transformationMatrix);
         UnpremultiplyBitmap(bitmap);
